frontend/requirements.txt
streamlit
openai
duckduckgo-search
matplotlib
numpy
scipy
pandas
pillow
requests
streamlit-audiorecorder
speech-recognition
pydub
sqlalchemy